fix from @jonboulle's comments
